third_party/blink/web_tests/fast/forms/time-multiple-fields/time-multiple-fields-value-set-empty-expected.txt

Check if input.value="" clears an input with partially-specified value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Empty text: --:-- --
PASS eventSender.keyDown("ArrowUp"); getUserAgentShadowTextContent(input) is not emptyText
PASS input.value = ""; getUserAgentShadowTextContent(input) is emptyText


Initial text: --:--:45 --
Empty value should not clear read-only fields.
FAIL input.value = ""; getUserAgentShadowTextContent(input) should be --:--:45 --. Was --:--:45 --.
Empty value should clear only editable fields.
FAIL eventSender.keyDown("ArrowUp"); getUserAgentShadowTextContent(input) should be 01:--:45 --. Was 01:--:45 --.
FAIL input.value = ""; getUserAgentShadowTextContent(input) should be --:--:45 --. Was --:--:45 --.
PASS successfullyParsed is true

TEST COMPLETE
